+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+- First release with the following features:
+  - Custom list item template
+  - Multiselect
+  - Single select
+  - Reorder
+  - Drag & Drop